refactor(Remove all Custom Repos and replace with default ones): Cusomt repos are due to deprecation
fix #63
