+++ __tests__/routes/users/users.delete.integration.test.ts
@@ -3,8 +3,7 @@ jest.useFakeTimers();
 import express, { Application } from 'express';
 import 'reflect-metadata';
 import request from 'supertest';
-import { Connection, getCustomRepository, getRepository } from 'typeorm';
-import { UserRepository } from '../../../src/lib/repositories/UserRepository';
+import { Connection, getRepository } from 'typeorm';
 import routes from '../../../src/lib/routes';
 import { User } from '../../../src/orm/entity/User';
 import {
@@ -36,14 +35,7 @@ describe('testing delete users', () => {
     connections = await createTestingConnections();
     done();
   });
-  // beforeEach(async (done) => {
-  //   try {
-  //     await reloadTestingDatabases(connections);
-  //     done();
-  //   } catch (err) {
-  //     console.warn(err.message);
-  //   }
-  // });
+
   afterAll(async (done) => {
     try {
       await reloadTestingDatabases(connections);
@@ -75,8 +67,6 @@ describe('testing delete users', () => {
 // ╚═════╝  ╚═════╝ ╚═╝  ╚═══╝╚══════╝
 
   test('should delete a users', async (done) => {
-  // process.env.NODE_ENV = 'development';
-  // const usersres = await request(app).get('/api/v1/users');
   const userRepo = getRepository(User);
   await userRepo.insert({
     email: 'foo@bah.com',
@@ -86,7 +76,6 @@ describe('testing delete users', () => {
     role: UserRole.creator,
  });
   const user = await userRepo.findOne({where: {firstName: 'Jimmy'}});
-  // console.log(users);
   const id = user.id;
 
   const res = await request(app).delete(`/api/v1/users/${id}`).set(headers);
@@ -107,8 +96,7 @@ describe('testing delete users', () => {
 });
 
   test('should delete user even if he has spots', async (done) => {
-  // process.env.NODE_ENV = 'development';
-  const userRepo = getCustomRepository(UserRepository);
+  const userRepo = getRepository(User);
   const usersWithRelations = await userRepo.find({
     relations: ['bathingspots'],
     where: {protected: false},
@@ -117,26 +105,20 @@ describe('testing delete users', () => {
   .delete(
     `/api/v1/users/${usersWithRelations[0].id}`)
     .set(headers);
-
-  // console.log(res.body);
   expect(res.status).toBe(200);
   expect(res.body.success).toBe(true);
-  // expect(res.body.data.length).toBe(0);
   done();
 });
   test('should fail. Can\'t delete protected user', async (done) => {
-  const userRepo = getCustomRepository(UserRepository);
+  const userRepo = getRepository(User);
   const usersWithRelations = await userRepo.find({relations: ['bathingspots'], where: {protected: true}});
 
-  // console.log(usersWithRelations);
   const res = await request(app)
   .delete(
     `/api/v1/users/${usersWithRelations[0].id}`)
     .set(headers);
   expect(res.status).toBe(403);
-  // console.log(res.body);
   expect(res.body.success).toBe(false);
-  // expect(res.body.data.length).toBe(0);
   done();
 });
 });
