+++ __tests__/routes/users/users.put.integration.test.ts
@@ -2,8 +2,7 @@ jest.useFakeTimers();
 import express, { Application } from 'express';
 import 'reflect-metadata';
 import request from 'supertest';
-import { Connection, getCustomRepository } from 'typeorm';
-import { UserRepository } from '../../../src/lib/repositories/UserRepository';
+import { Connection } from 'typeorm';
 import routes from '../../../src/lib/routes';
 import { DefaultRegions, UserRole } from '../../../src/lib/common';
 import {
@@ -13,6 +12,7 @@ import {
   readTokenFromDisc,
 } from '../../test-utils';
 import path from 'path';
+import { User } from '../../../src/orm/entity';
 // ███████╗███████╗████████╗██╗   ██╗██████╗
 // ██╔════╝██╔════╝╚══██╔══╝██║   ██║██╔══██╗
 // ███████╗█████╗     ██║   ██║   ██║██████╔╝
@@ -74,10 +74,8 @@ describe('testing put users', () => {
 // ╚═════╝  ╚═════╝ ╚═╝  ╚═══╝╚══════╝
 
   test('update user', async (done) => {
-  // process.env.NODE_ENV = 'development';
-  // expect.assertions(2);
+
   const usersres = await request(app).get('/api/v1/users').set(headers);
-  // console.log(usersres);
   const id = usersres.body.data[usersres.body.data.length - 1].id;
   const res = await request(app).put(`/api/v1/users/${id}`).send({
     email: 'foo@test.com',
@@ -98,9 +96,7 @@ describe('testing put users', () => {
     role: UserRole.creator,
   })
     .set(headers);
-  // console.log(newUserRes.body);
-  // const usersres = await request(app).get('/api/v1/users');
-  // const id = usersres.body.data[usersres.body.data.length - 1].id;
+
   await request(app).post(`/api/v1/users/${newUserRes.body.data[0].id}/bathingspots`).send({
     isPublic: false,
     name: 'intermidiante spot',
@@ -120,30 +116,20 @@ describe('testing put users', () => {
   done();
 });
 
-  test('user fail due to undefiend user id', async (done) => {
-  const userRepo = getCustomRepository(UserRepository);
-  // const usersWithRelations = await userRepo.find({relations: ['bathingspots']});
+  test('user fail due to undefined user id', async (done) => {
 
-  // console.log(usersWithRelations);
   const res = await request(app).put(
     `/api/v1/users/${1000}`).set(headers);
   expect(res.status).toBe(404);
-  // console.log(res.body);
   expect(res.body.success).toBe(false);
-  // expect(res.body.data.length).toBe(0);
   done();
 });
   test('user fail due to wrong route', async (done) => {
-  const userRepo = getCustomRepository(UserRepository);
-  // const usersWithRelations = await userRepo.find({relations: ['bathingspots']});
 
-  // console.log(usersWithRelations);
   const res = await request(app).put(
     `/api/v1/users/`).set(headers);
   expect(res.status).toBe(404);
-  // console.log(res.body);
-  // expect(res.body.success).toBe(false);
-  // expect(res.body.data.length).toBe(0);
+
   done();
 });
 });
