+++ src/lib/request-handlers/bathingspots/public-get.ts
@@ -1,11 +1,10 @@
-import { getCustomRepository, getRepository } from 'typeorm';
+import { getRepository } from 'typeorm';
 import { Bathingspot } from '../../../orm/entity/Bathingspot';
 import { SUCCESS } from '../../messages';
-import { RegionRepository } from '../../repositories/RegionRepository';
 import { getResponse, HttpCodes } from '../../common';
 import { errorResponse, responder, responderWrongId, successResponse } from '../responders';
 import { findByRegionId } from '../../utils/spot-repo-helpers';
-import { getRegionsList } from '../../utils/region-repo-helpers';
+import { getRegionsList, findByName } from '../../utils/region-repo-helpers';
 
 /**
  * Todo: Which properties should be returned
@@ -53,9 +52,8 @@ export const getBathingspotsByRegion: getResponse = async (request, response) =>
     if (!(list.includes(request.params.region))) {
       responderWrongId(response);
     } else {
-      const regionRepo = getCustomRepository(RegionRepository);
       // const spotRepo = getCustomRepository(BathingspotRepository);
-      const region = await regionRepo.findByName(request.params.region);
+      const region = await findByName(request.params.region);
       let spots: []|any = [];
       if (region !== undefined) {
         spots = await findByRegionId(region.id);
