+++ src/lib/request-handlers/regions/delete.ts
@@ -1,15 +1,16 @@
-import { getCustomRepository } from 'typeorm';
-import { RegionRepository } from '../../repositories/RegionRepository';
 import { deleteResponse, HttpCodes } from '../../common';
 import { errorResponse, responder, responderWrongId, successResponse } from '../responders';
 
 import { SUCCESS } from '../../messages';
+import { findByIdWithRelations } from '../../utils/region-repo-helpers';
+import { getRepository } from 'typeorm';
+import { Region } from '../../../orm/entity/Region';
 
 export const deleteRegion: deleteResponse = async (request, response) => {
   try {
-    const regionRepo = getCustomRepository(RegionRepository);
+    const regionRepo = getRepository(Region);
     const regionId = request.params.regionId;
-    const region = await regionRepo.findByIdWithRelations(regionId, ['bathingspots']);
+    const region = await findByIdWithRelations(regionId, ['bathingspots']);
     if (region === undefined) {
       responderWrongId(response);
     } else {
