+++ src/setup/import-existing-data.ts
@@ -3,7 +3,7 @@ import { extname, resolve } from 'path';
 import proj4 from 'proj4';
 import { promisify } from 'util';
 import { IObject } from '../lib/common';
-import { createSpotWithValues } from '../lib/utils/bathingspot-helpers';
+import { createSpotWithValues } from '../lib/utils/spot-helpers';
 import { createMeasurementWithValues } from '../lib/utils/measurement-helpers';
 import { Bathingspot } from '../orm/entity/Bathingspot';
 import { createPredictionWithValues } from '../lib/utils/predictions-helpers';
